Removed unused Utils file
